requirements.txt
backtrader[plotting]==1.9.78.123
matplotlib==3.10.3
numpy_financial==1.0.0
thematicnifty==0.1.4
requests==2.32.4
yfinance==0.2.65
numpy_financial==1.0.0